Bugfixes
- Now ChangeItemIconTransaction does not delete "no_icon.png"
- Now the map button has no margin at the left
- Now the manifest sets the right origin of the application

+++ API/class/transactions/ChangeItemIconTransaction.php
@@ -47,7 +47,7 @@ class ChangeItemIconTransaction extends Transaction {
     }
 
     private function deleteIfExists($icon) {
-        if(isset($icon) && trim($icon) !== '') {
+        if(isset($icon) && trim($icon) !== '' && trim($icon) !== 'no_icon.png') {
             $iconFile = \IApiInterface::ITEM_ICONS_DIR.$icon;
             if(file_exists($iconFile)) {
                 unlink($iconFile);
